+++ CHANGELOG.md
@@ -7,6 +7,24 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Development]
 
+## [0.25.0 - 2022-05-01]
+
+Major version bump due to breaking igraph change
+
+### Added
+
+* igraph handlers: `graphistry.from_igraph`, `g.from_igraph`, `g.to_igraph`
+* docs: README.md examples of using new igraph methods
+
+### Changed
+
+* Deprecation warnings in old igraph methods: `g.graph(ig)`, `igraph2pandas`, `pandas2igraph`
+* Internal igraph handlers upgraded to use new igraph methods 
+
+### Breaking
+
+* `network2igraph` and `igraph2pandas` renamed output node ID column to `_n_implicit` (`constants.NODE`)
+
 ## [0.24.1 - 2022-04-29]
 
 ### Fixed
@@ -14,6 +32,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Expose symbols for `.chain()` predicates as top-level: previous `ast` export was incorrect
 
 ## [0.24.0 - 2022-04-29]
+
+Major version bump due to large dependency increases for kitchen-sink installs and overall sizeable new feature
 
 ### Added